requirements.txt
requests
python-telegram-bot[job-queue]>=21.0